fix(ta_codegen): widen float inputs to double in single-precision variants (#33)
The generated single-precision TA_S_* functions did their intermediate
arithmetic in float before widening to the double output, so combining two
inputs could overflow to infinity (e.g. TA_S_MULT of two large floats) or
lose precision. The C and Java backends now widen every float input read to
double when emitting a TA_S_ variant, so TA_S_<F>(float) is bit-identical to
TA_<F>((double)float).

Adds a single-precision overflow regression per language (ta_regtest
test_func_s_overflow and Java TestSMathOverflow) and retargets the codegen
float leg to assert the TA_S_ == double-on-widened invariant bit-for-bit.

+++ src/ta_func/ta_ACCBANDS.c
@@ -377,16 +377,16 @@ TA_RetCode TA_S_ACCBANDS( int    startIdx,
    }
    for( j = 0, i = startIdx - lookbackTotal; i <= endIdx; i += 1, j += 1 )
    {
-      tempReal = inHigh[i] + inLow[i];
+      tempReal = (double)inHigh[i] + (double)inLow[i];
       if( !TA_IS_ZERO(tempReal) )
       {
-         tempReal = 4 * (inHigh[i] - inLow[i]) / tempReal;
-         tempBuffer1[j] = inHigh[i] * (1 + tempReal);
-         tempBuffer2[j] = inLow[i] * (1 - tempReal);
+         tempReal = 4 * ((double)inHigh[i] - (double)inLow[i]) / tempReal;
+         tempBuffer1[j] = (double)inHigh[i] * (1 + tempReal);
+         tempBuffer2[j] = (double)inLow[i] * (1 - tempReal);
       } else 
       {
-         tempBuffer1[j] = inHigh[i];
-         tempBuffer2[j] = inLow[i];
+         tempBuffer1[j] = (double)inHigh[i];
+         tempBuffer2[j] = (double)inLow[i];
       }
    }
    retCode = TA_S_SMA_Unguarded(startIdx,endIdx,inClose,optInTimePeriod,&outBegIdxDummy,&outNbElementDummy,outRealMiddleBand);
@@ -475,16 +475,16 @@ TA_RetCode TA_S_ACCBANDS_Unguarded( int    startIdx,
    }
    for( j = 0, i = startIdx - lookbackTotal; i <= endIdx; i += 1, j += 1 )
    {
-      tempReal = inHigh[i] + inLow[i];
+      tempReal = (double)inHigh[i] + (double)inLow[i];
       if( !TA_IS_ZERO(tempReal) )
       {
-         tempReal = 4 * (inHigh[i] - inLow[i]) / tempReal;
-         tempBuffer1[j] = inHigh[i] * (1 + tempReal);
-         tempBuffer2[j] = inLow[i] * (1 - tempReal);
+         tempReal = 4 * ((double)inHigh[i] - (double)inLow[i]) / tempReal;
+         tempBuffer1[j] = (double)inHigh[i] * (1 + tempReal);
+         tempBuffer2[j] = (double)inLow[i] * (1 - tempReal);
       } else 
       {
-         tempBuffer1[j] = inHigh[i];
-         tempBuffer2[j] = inLow[i];
+         tempBuffer1[j] = (double)inHigh[i];
+         tempBuffer2[j] = (double)inLow[i];
       }
    }
    retCode = TA_S_SMA_Unguarded(startIdx,endIdx,inClose,optInTimePeriod,&outBegIdxDummy,&outNbElementDummy,outRealMiddleBand);
