+++ src/ta_func/ta_ADD.c
@@ -137,7 +137,7 @@ TA_RetCode TA_S_ADD( int    startIdx,
 
    for( i = startIdx, outIdx = 0; i <= endIdx; i += 1, outIdx += 1 )
    {
-      outReal[outIdx] = inReal0[i] + inReal1[i];
+      outReal[outIdx] = (double)inReal0[i] + (double)inReal1[i];
    }
    *outNBElement= outIdx;
    *outBegIdx= startIdx;
@@ -157,7 +157,7 @@ TA_RetCode TA_S_ADD_Unguarded( int    startIdx,
 
    for( i = startIdx, outIdx = 0; i <= endIdx; i += 1, outIdx += 1 )
    {
-      outReal[outIdx] = inReal0[i] + inReal1[i];
+      outReal[outIdx] = (double)inReal0[i] + (double)inReal1[i];
    }
    *outNBElement= outIdx;
    *outBegIdx= startIdx;
