+++ src/ta_func/ta_ADOSC.c
@@ -418,10 +418,10 @@ TA_RetCode TA_S_ADOSC( int    startIdx,
    one_minus_fastk = 1.0 - fastk;
    slowk = 2.0 / ((double)optInSlowPeriod + 1.0);
    one_minus_slowk = 1.0 - slowk;
-   high = inHigh[today];
-   low = inLow[today];
+   high = (double)inHigh[today];
+   low = (double)inLow[today];
    tmp = high - low;
-   close = inClose[today];
+   close = (double)inClose[today];
    if( tmp > 0.0 )
    {
       ad += (close - low - (high - close)) / tmp * (double)inVolume[today];
@@ -431,10 +431,10 @@ TA_RetCode TA_S_ADOSC( int    startIdx,
    slowEMA = ad;
    while( today < startIdx )
    {
-      high = inHigh[today];
-      low = inLow[today];
+      high = (double)inHigh[today];
+      low = (double)inLow[today];
       tmp = high - low;
-      close = inClose[today];
+      close = (double)inClose[today];
       if( tmp > 0.0 )
       {
          ad += (close - low - (high - close)) / tmp * (double)inVolume[today];
@@ -446,10 +446,10 @@ TA_RetCode TA_S_ADOSC( int    startIdx,
    outIdx = 0;
    while( today <= endIdx )
    {
-      high = inHigh[today];
-      low = inLow[today];
+      high = (double)inHigh[today];
+      low = (double)inLow[today];
       tmp = high - low;
-      close = inClose[today];
+      close = (double)inClose[today];
       if( tmp > 0.0 )
       {
          ad += (close - low - (high - close)) / tmp * (double)inVolume[today];
@@ -516,10 +516,10 @@ TA_RetCode TA_S_ADOSC_Unguarded( int    startIdx,
    one_minus_fastk = 1.0 - fastk;
    slowk = 2.0 / ((double)optInSlowPeriod + 1.0);
    one_minus_slowk = 1.0 - slowk;
-   high = inHigh[today];
-   low = inLow[today];
+   high = (double)inHigh[today];
+   low = (double)inLow[today];
    tmp = high - low;
-   close = inClose[today];
+   close = (double)inClose[today];
    if( tmp > 0.0 )
    {
       ad += (close - low - (high - close)) / tmp * (double)inVolume[today];
@@ -529,10 +529,10 @@ TA_RetCode TA_S_ADOSC_Unguarded( int    startIdx,
    slowEMA = ad;
    while( today < startIdx )
    {
-      high = inHigh[today];
-      low = inLow[today];
+      high = (double)inHigh[today];
+      low = (double)inLow[today];
       tmp = high - low;
-      close = inClose[today];
+      close = (double)inClose[today];
       if( tmp > 0.0 )
       {
          ad += (close - low - (high - close)) / tmp * (double)inVolume[today];
@@ -544,10 +544,10 @@ TA_RetCode TA_S_ADOSC_Unguarded( int    startIdx,
    outIdx = 0;
    while( today <= endIdx )
    {
-      high = inHigh[today];
-      low = inLow[today];
+      high = (double)inHigh[today];
+      low = (double)inLow[today];
       tmp = high - low;
-      close = inClose[today];
+      close = (double)inClose[today];
       if( tmp > 0.0 )
       {
          ad += (close - low - (high - close)) / tmp * (double)inVolume[today];
