+++ src/ta_func/ta_ADX.c
@@ -772,17 +772,17 @@ TA_RetCode TA_S_ADX( int    startIdx,
    prevPlusDM = 0.0;
    prevTR = 0.0;
    today = startIdx - lookbackTotal;
-   prevHigh = inHigh[today];
-   prevLow = inLow[today];
-   prevClose = inClose[today];
+   prevHigh = (double)inHigh[today];
+   prevLow = (double)inLow[today];
+   prevClose = (double)inClose[today];
    i = optInTimePeriod - 1;
    while( i-- > 0 )
    {
       today += 1;
-      tempReal = inHigh[today];
+      tempReal = (double)inHigh[today];
       diffP = tempReal - prevHigh;
       prevHigh = tempReal;
-      tempReal = inLow[today];
+      tempReal = (double)inLow[today];
       diffM = prevLow - tempReal;
       prevLow = tempReal;
       if( diffM > 0 && diffP < diffM )
@@ -807,17 +807,17 @@ TA_RetCode TA_S_ADX( int    startIdx,
       _true_range_0 = range_0;
       tempReal = _true_range_0;
       prevTR += tempReal;
-      prevClose = inClose[today];
+      prevClose = (double)inClose[today];
    }
    sumDX = 0.0;
    i = optInTimePeriod;
    while( i-- > 0 )
    {
       today += 1;
-      tempReal = inHigh[today];
+      tempReal = (double)inHigh[today];
       diffP = tempReal - prevHigh;
       prevHigh = tempReal;
-      tempReal = inLow[today];
+      tempReal = (double)inLow[today];
       diffM = prevLow - tempReal;
       prevLow = tempReal;
       prevMinusDM -= prevMinusDM / optInTimePeriod;
@@ -844,7 +844,7 @@ TA_RetCode TA_S_ADX( int    startIdx,
       _true_range_1 = range_1;
       tempReal = _true_range_1;
       prevTR = prevTR - prevTR / optInTimePeriod + tempReal;
-      prevClose = inClose[today];
+      prevClose = (double)inClose[today];
       if( !TA_IS_ZERO(prevTR) )
       {
          minusDI = (100.0 * (prevMinusDM / prevTR));
@@ -861,10 +861,10 @@ TA_RetCode TA_S_ADX( int    startIdx,
    while( i-- > 0 )
    {
       today += 1;
-      tempReal = inHigh[today];
+      tempReal = (double)inHigh[today];
       diffP = tempReal - prevHigh;
       prevHigh = tempReal;
-      tempReal = inLow[today];
+      tempReal = (double)inLow[today];
       diffM = prevLow - tempReal;
       prevLow = tempReal;
       prevMinusDM -= prevMinusDM / optInTimePeriod;
@@ -891,7 +891,7 @@ TA_RetCode TA_S_ADX( int    startIdx,
       _true_range_2 = range_2;
       tempReal = _true_range_2;
       prevTR = prevTR - prevTR / optInTimePeriod + tempReal;
-      prevClose = inClose[today];
+      prevClose = (double)inClose[today];
       if( !TA_IS_ZERO(prevTR) )
       {
          minusDI = (100.0 * (prevMinusDM / prevTR));
@@ -909,10 +909,10 @@ TA_RetCode TA_S_ADX( int    startIdx,
    while( today < endIdx )
    {
       today += 1;
-      tempReal = inHigh[today];
+      tempReal = (double)inHigh[today];
       diffP = tempReal - prevHigh;
       prevHigh = tempReal;
-      tempReal = inLow[today];
+      tempReal = (double)inLow[today];
       diffM = prevLow - tempReal;
       prevLow = tempReal;
       prevMinusDM -= prevMinusDM / optInTimePeriod;
@@ -939,7 +939,7 @@ TA_RetCode TA_S_ADX( int    startIdx,
       _true_range_3 = range_3;
       tempReal = _true_range_3;
       prevTR = prevTR - prevTR / optInTimePeriod + tempReal;
-      prevClose = inClose[today];
+      prevClose = (double)inClose[today];
       if( !TA_IS_ZERO(prevTR) )
       {
          minusDI = (100.0 * (prevMinusDM / prevTR));
@@ -1004,17 +1004,17 @@ TA_RetCode TA_S_ADX_Unguarded( int    startIdx,
    prevPlusDM = 0.0;
    prevTR = 0.0;
    today = startIdx - lookbackTotal;
-   prevHigh = inHigh[today];
-   prevLow = inLow[today];
-   prevClose = inClose[today];
+   prevHigh = (double)inHigh[today];
+   prevLow = (double)inLow[today];
+   prevClose = (double)inClose[today];
    i = optInTimePeriod - 1;
    while( i-- > 0 )
    {
       today += 1;
-      tempReal = inHigh[today];
+      tempReal = (double)inHigh[today];
       diffP = tempReal - prevHigh;
       prevHigh = tempReal;
-      tempReal = inLow[today];
+      tempReal = (double)inLow[today];
       diffM = prevLow - tempReal;
       prevLow = tempReal;
       if( diffM > 0 && diffP < diffM )
@@ -1039,17 +1039,17 @@ TA_RetCode TA_S_ADX_Unguarded( int    startIdx,
       _true_range_0 = range_0;
       tempReal = _true_range_0;
       prevTR += tempReal;
-      prevClose = inClose[today];
+      prevClose = (double)inClose[today];
    }
    sumDX = 0.0;
    i = optInTimePeriod;
    while( i-- > 0 )
    {
       today += 1;
-      tempReal = inHigh[today];
+      tempReal = (double)inHigh[today];
       diffP = tempReal - prevHigh;
       prevHigh = tempReal;
-      tempReal = inLow[today];
+      tempReal = (double)inLow[today];
       diffM = prevLow - tempReal;
       prevLow = tempReal;
       prevMinusDM -= prevMinusDM / optInTimePeriod;
@@ -1076,7 +1076,7 @@ TA_RetCode TA_S_ADX_Unguarded( int    startIdx,
       _true_range_1 = range_1;
       tempReal = _true_range_1;
       prevTR = prevTR - prevTR / optInTimePeriod + tempReal;
-      prevClose = inClose[today];
+      prevClose = (double)inClose[today];
       if( !TA_IS_ZERO(prevTR) )
       {
          minusDI = (100.0 * (prevMinusDM / prevTR));
@@ -1093,10 +1093,10 @@ TA_RetCode TA_S_ADX_Unguarded( int    startIdx,
    while( i-- > 0 )
    {
       today += 1;
-      tempReal = inHigh[today];
+      tempReal = (double)inHigh[today];
       diffP = tempReal - prevHigh;
       prevHigh = tempReal;
-      tempReal = inLow[today];
+      tempReal = (double)inLow[today];
       diffM = prevLow - tempReal;
       prevLow = tempReal;
       prevMinusDM -= prevMinusDM / optInTimePeriod;
@@ -1123,7 +1123,7 @@ TA_RetCode TA_S_ADX_Unguarded( int    startIdx,
       _true_range_2 = range_2;
       tempReal = _true_range_2;
       prevTR = prevTR - prevTR / optInTimePeriod + tempReal;
-      prevClose = inClose[today];
+      prevClose = (double)inClose[today];
       if( !TA_IS_ZERO(prevTR) )
       {
          minusDI = (100.0 * (prevMinusDM / prevTR));
@@ -1141,10 +1141,10 @@ TA_RetCode TA_S_ADX_Unguarded( int    startIdx,
    while( today < endIdx )
    {
       today += 1;
-      tempReal = inHigh[today];
+      tempReal = (double)inHigh[today];
       diffP = tempReal - prevHigh;
       prevHigh = tempReal;
-      tempReal = inLow[today];
+      tempReal = (double)inLow[today];
       diffM = prevLow - tempReal;
       prevLow = tempReal;
       prevMinusDM -= prevMinusDM / optInTimePeriod;
@@ -1171,7 +1171,7 @@ TA_RetCode TA_S_ADX_Unguarded( int    startIdx,
       _true_range_3 = range_3;
       tempReal = _true_range_3;
       prevTR = prevTR - prevTR / optInTimePeriod + tempReal;
-      prevClose = inClose[today];
+      prevClose = (double)inClose[today];
       if( !TA_IS_ZERO(prevTR) )
       {
          minusDI = (100.0 * (prevMinusDM / prevTR));
