+++ src/ta_func/ta_AROONOSC.c
@@ -361,15 +361,15 @@ TA_RetCode TA_S_AROONOSC( int    startIdx,
    factor = (double)100.0 / (double)optInTimePeriod;
    while( today <= endIdx )
    {
-      tmp = inLow[today];
+      tmp = (double)inLow[today];
       if( lowestIdx < trailingIdx )
       {
          lowestIdx = trailingIdx;
-         lowest = inLow[lowestIdx];
+         lowest = (double)inLow[lowestIdx];
          i = lowestIdx;
          while( ++i <= today )
          {
-            tmp = inLow[i];
+            tmp = (double)inLow[i];
             if( tmp <= lowest )
             {
                lowestIdx = i;
@@ -381,15 +381,15 @@ TA_RetCode TA_S_AROONOSC( int    startIdx,
          lowestIdx = today;
          lowest = tmp;
       }
-      tmp = inHigh[today];
+      tmp = (double)inHigh[today];
       if( highestIdx < trailingIdx )
       {
          highestIdx = trailingIdx;
-         highest = inHigh[highestIdx];
+         highest = (double)inHigh[highestIdx];
          i = highestIdx;
          while( ++i <= today )
          {
-            tmp = inHigh[i];
+            tmp = (double)inHigh[i];
             if( tmp >= highest )
             {
                highestIdx = i;
@@ -453,15 +453,15 @@ TA_RetCode TA_S_AROONOSC_Unguarded( int    startIdx,
    factor = (double)100.0 / (double)optInTimePeriod;
    while( today <= endIdx )
    {
-      tmp = inLow[today];
+      tmp = (double)inLow[today];
       if( lowestIdx < trailingIdx )
       {
          lowestIdx = trailingIdx;
-         lowest = inLow[lowestIdx];
+         lowest = (double)inLow[lowestIdx];
          i = lowestIdx;
          while( ++i <= today )
          {
-            tmp = inLow[i];
+            tmp = (double)inLow[i];
             if( tmp <= lowest )
             {
                lowestIdx = i;
@@ -473,15 +473,15 @@ TA_RetCode TA_S_AROONOSC_Unguarded( int    startIdx,
          lowestIdx = today;
          lowest = tmp;
       }
-      tmp = inHigh[today];
+      tmp = (double)inHigh[today];
       if( highestIdx < trailingIdx )
       {
          highestIdx = trailingIdx;
-         highest = inHigh[highestIdx];
+         highest = (double)inHigh[highestIdx];
          i = highestIdx;
          while( ++i <= today )
          {
-            tmp = inHigh[i];
+            tmp = (double)inHigh[i];
             if( tmp >= highest )
             {
                highestIdx = i;
