+++ src/ta_func/ta_ATR.c
@@ -440,9 +440,9 @@ TA_RetCode TA_S_ATR( int    startIdx,
    i = optInTimePeriod;
    while( i-- > 0 )
    {
-      tempLT = inLow[today];
-      tempHT = inHigh[today];
-      tempCY = inClose[today - 1];
+      tempLT = (double)inLow[today];
+      tempHT = (double)inHigh[today];
+      tempCY = (double)inClose[today - 1];
       greatest = tempHT - tempLT;
       val2 = fabs(tempCY - tempHT);
       if( val2 > greatest )
@@ -461,9 +461,9 @@ TA_RetCode TA_S_ATR( int    startIdx,
    i = TA_GLOBALS_UNSTABLE_PERIOD(TA_FUNC_UNST_ATR,Atr);
    while( i != 0 )
    {
-      tempLT = inLow[today];
-      tempHT = inHigh[today];
-      tempCY = inClose[today - 1];
+      tempLT = (double)inLow[today];
+      tempHT = (double)inHigh[today];
+      tempCY = (double)inClose[today - 1];
       greatest = tempHT - tempLT;
       val2 = fabs(tempCY - tempHT);
       if( val2 > greatest )
@@ -486,9 +486,9 @@ TA_RetCode TA_S_ATR( int    startIdx,
    nbATR = endIdx - startIdx + 1;
    while( --nbATR != 0 )
    {
-      tempLT = inLow[today];
-      tempHT = inHigh[today];
-      tempCY = inClose[today - 1];
+      tempLT = (double)inLow[today];
+      tempHT = (double)inHigh[today];
+      tempCY = (double)inClose[today - 1];
       greatest = tempHT - tempLT;
       val2 = fabs(tempCY - tempHT);
       if( val2 > greatest )
@@ -555,9 +555,9 @@ TA_RetCode TA_S_ATR_Unguarded( int    startIdx,
    i = optInTimePeriod;
    while( i-- > 0 )
    {
-      tempLT = inLow[today];
-      tempHT = inHigh[today];
-      tempCY = inClose[today - 1];
+      tempLT = (double)inLow[today];
+      tempHT = (double)inHigh[today];
+      tempCY = (double)inClose[today - 1];
       greatest = tempHT - tempLT;
       val2 = fabs(tempCY - tempHT);
       if( val2 > greatest )
@@ -576,9 +576,9 @@ TA_RetCode TA_S_ATR_Unguarded( int    startIdx,
    i = TA_GLOBALS_UNSTABLE_PERIOD(TA_FUNC_UNST_ATR,Atr);
    while( i != 0 )
    {
-      tempLT = inLow[today];
-      tempHT = inHigh[today];
-      tempCY = inClose[today - 1];
+      tempLT = (double)inLow[today];
+      tempHT = (double)inHigh[today];
+      tempCY = (double)inClose[today - 1];
       greatest = tempHT - tempLT;
       val2 = fabs(tempCY - tempHT);
       if( val2 > greatest )
@@ -601,9 +601,9 @@ TA_RetCode TA_S_ATR_Unguarded( int    startIdx,
    nbATR = endIdx - startIdx + 1;
    while( --nbATR != 0 )
    {
-      tempLT = inLow[today];
-      tempHT = inHigh[today];
-      tempCY = inClose[today - 1];
+      tempLT = (double)inLow[today];
+      tempHT = (double)inHigh[today];
+      tempCY = (double)inClose[today - 1];
       greatest = tempHT - tempLT;
       val2 = fabs(tempCY - tempHT);
       if( val2 > greatest )
